+++ cryptofeed/exchanges/ftx.py
@@ -398,7 +398,7 @@ class FTX(Feed, FTXRestMixin):
             BUY if order['side'].lower() == 'buy' else SELL,
             status,
             LIMIT if order['type'] == 'limit' else MARKET,
-            Decimal(order['price']),
+            Decimal(order['price']) if order['price'] else None,
             Decimal(order['filledSize']),
             Decimal(order['remainingSize']),
             None,